refactoring, split to components
